packages/quality/python/pyproject.toml
[project]
name = "rsf-quality-python"
version = "1.0.0"
description = "Layer 4 Python — DSPy evaluations, Hypothesis property-based testing"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.30.0",
    "dspy>=2.5.0",
    "hypothesis>=6.100.0",
    "ollama>=0.3.0",
    "psycopg2-binary>=2.9.9",
    "pytest>=8.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.10.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
include = ["evaluations*", "pbt*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
strict = true
python_version = "3.11"
